main.py
import os
import pickle
from typing import Optional, List, Dict, Any, Tuple
from difflib import get_close_matches

import numpy as np
import pandas as pd
import httpx
from fastapi import FastAPI, HTTPException, Query
from fastapi.middleware.cors import CORSMiddleware
from pydantic import BaseModel, Field
from dotenv import load_dotenv

# =========================
# ENV
# =========================
load_dotenv()
TMDB_API_KEY = os.getenv("TMDB_API_KEY")

TMDB_BASE = "https://api.themoviedb.org/3"
TMDB_IMG_500 = "https://image.tmdb.org/t/p/w500"

if not TMDB_API_KEY:
    raise RuntimeError("TMDB_API_KEY missing. Put it in .env as TMDB_API_KEY=xxxx")

# =========================
# FASTAPI APP
# =========================
app = FastAPI(title="Movie Recommender API", version="3.1")

app.add_middleware(
    CORSMiddleware,
    allow_origins=["*"],
    allow_credentials=True,
    allow_methods=["*"],
    allow_headers=["*"],
)

# =========================
# PICKLE GLOBALS
# =========================
BASE_DIR = os.path.dirname(os.path.abspath(__file__))

DF_PATH = os.path.join(BASE_DIR, "df.pkl")
INDICES_PATH = os.path.join(BASE_DIR, "indices.pkl")
TFIDF_MATRIX_PATH = os.path.join(BASE_DIR, "tfidf_matrix.pkl")
TFIDF_PATH = os.path.join(BASE_DIR, "tfidf.pkl")

df: Optional[pd.DataFrame] = None
tfidf_matrix: Any = None
TITLE_TO_IDX: Optional[Dict[str, int]] = None

# =========================
# MODELS
# =========================
class TMDBMovieCard(BaseModel):
    tmdb_id: int
    title: str
    poster_url: Optional[str] = None
    release_date: Optional[str] = None
    vote_average: Optional[float] = None


class TMDBMovieDetails(BaseModel):
    tmdb_id: int
    title: str
    overview: Optional[str] = None
    release_date: Optional[str] = None
    poster_url: Optional[str] = None
    backdrop_url: Optional[str] = None
    genres: List[dict] = Field(default_factory=list)


class TFIDFRecItem(BaseModel):
    title: str
    score: float
    tmdb: Optional[TMDBMovieCard] = None


class SearchBundleResponse(BaseModel):
    query: str
    movie_details: TMDBMovieDetails
    tfidf_recommendations: List[TFIDFRecItem]
    genre_recommendations: List[TMDBMovieCard]

# =========================
# UTILS
# =========================
def _norm_title(t: str) -> str:
    return str(t).strip().lower()


def make_img_url(path: Optional[str]) -> Optional[str]:
    return f"{TMDB_IMG_500}{path}" if path else None


async def tmdb_get(path: str, params: Dict[str, Any]) -> Dict[str, Any]:
    q = dict(params)
    q["api_key"] = TMDB_API_KEY

    async with httpx.AsyncClient(timeout=20) as client:
        r = await client.get(f"{TMDB_BASE}{path}", params=q)

    if r.status_code != 200:
        raise HTTPException(status_code=502, detail=f"TMDB error {r.status_code}")

    return r.json()


# =========================
# TMDB HELPERS
# =========================
async def tmdb_cards_from_results(results: List[dict], limit: int):
    return [
        TMDBMovieCard(
            tmdb_id=int(m["id"]),
            title=m.get("title") or "",
            poster_url=make_img_url(m.get("poster_path")),
            release_date=m.get("release_date"),
            vote_average=m.get("vote_average"),
        )
        for m in (results or [])[:limit]
    ]


async def tmdb_movie_details(movie_id: int) -> TMDBMovieDetails:
    data = await tmdb_get(f"/movie/{movie_id}", {"language": "en-US"})
    return TMDBMovieDetails(
        tmdb_id=data["id"],
        title=data.get("title"),
        overview=data.get("overview"),
        release_date=data.get("release_date"),
        poster_url=make_img_url(data.get("poster_path")),
        backdrop_url=make_img_url(data.get("backdrop_path")),
        genres=data.get("genres", []),
    )


async def tmdb_search_first(query: str) -> Optional[dict]:
    data = await tmdb_get(
        "/search/movie",
        {"query": query, "language": "en-US", "include_adult": "false"},
    )
    return data["results"][0] if data.get("results") else None


# =========================
# TF-IDF HELPERS (FIXED)
# =========================
def build_title_to_idx_map(indices: Any) -> Dict[str, int]:
    return {_norm_title(k): int(v) for k, v in indices.items()}


def get_best_local_title(title: str) -> Optional[str]:
    matches = get_close_matches(
        _norm_title(title),
        TITLE_TO_IDX.keys(),
        n=1,
        cutoff=0.6,
    )
    return matches[0] if matches else None


def get_local_idx_by_title(title: str) -> int:
    key = _norm_title(title)
    if key in TITLE_TO_IDX:
        return TITLE_TO_IDX[key]

    fuzzy = get_best_local_title(title)
    if fuzzy:
        return TITLE_TO_IDX[fuzzy]

    raise HTTPException(status_code=404, detail="Title not in dataset")


def tfidf_recommend_titles(title: str, top_n: int):
    idx = get_local_idx_by_title(title)
    qv = tfidf_matrix[idx]
    scores = (tfidf_matrix @ qv.T).toarray().ravel()
    order = np.argsort(-scores)

    out = []
    for i in order:
        if i == idx:
            continue
        out.append((df.iloc[i]["title"], float(scores[i])))
        if len(out) >= top_n:
            break
    return out


async def attach_tmdb_card_by_title(title: str):
    try:
        m = await tmdb_search_first(title)
        if not m:
            return None
        return TMDBMovieCard(
            tmdb_id=m["id"],
            title=m.get("title"),
            poster_url=make_img_url(m.get("poster_path")),
            release_date=m.get("release_date"),
            vote_average=m.get("vote_average"),
        )
    except Exception:
        return None


# =========================
# STARTUP
# =========================
@app.on_event("startup")
def load_pickles():
    global df, tfidf_matrix, TITLE_TO_IDX

    with open(DF_PATH, "rb") as f:
        df = pickle.load(f)

    with open(INDICES_PATH, "rb") as f:
        TITLE_TO_IDX = build_title_to_idx_map(pickle.load(f))

    with open(TFIDF_MATRIX_PATH, "rb") as f:
        tfidf_matrix = pickle.load(f)

# =========================
# ROUTES
# =========================
@app.get("/home", response_model=List[TMDBMovieCard])
async def home(category: str = "popular", limit: int = 24):
    if category == "trending":
        data = await tmdb_get("/trending/movie/day", {"language": "en-US"})
    else:
        data = await tmdb_get(f"/movie/{category}", {"language": "en-US"})
    return await tmdb_cards_from_results(data.get("results", []), limit)


@app.get("/tmdb/search")
async def tmdb_search(query: str):
    return await tmdb_get(
        "/search/movie",
        {"query": query, "language": "en-US", "include_adult": "false"},
    )


@app.get("/movie/id/{tmdb_id}", response_model=TMDBMovieDetails)
async def movie_details_route(tmdb_id: int):
    return await tmdb_movie_details(tmdb_id)


@app.get("/movie/search", response_model=SearchBundleResponse)
async def search_bundle(query: str, tfidf_top_n: int = 12, genre_limit: int = 12):
    best = await tmdb_search_first(query)
    if not best:
        raise HTTPException(status_code=404, detail="Movie not found")

    details = await tmdb_movie_details(best["id"])

    tfidf_items = []
    try:
        recs = tfidf_recommend_titles(details.title, tfidf_top_n)
    except Exception:
        recs = []

    for t, s in recs:
        card = await attach_tmdb_card_by_title(t)
        tfidf_items.append(TFIDFRecItem(title=t, score=s, tmdb=card))

    genre_recs = []
    if details.genres:
        gid = details.genres[0]["id"]
        d = await tmdb_get(
            "/discover/movie",
            {"with_genres": gid, "sort_by": "popularity.desc"},
        )
        genre_recs = await tmdb_cards_from_results(d.get("results", []), genre_limit)

    # 🔥 GUARANTEED FALLBACK
    if not tfidf_items and genre_recs:
        tfidf_items = [
            TFIDFRecItem(title=c.title, score=0.0, tmdb=c)
            for c in genre_recs[:tfidf_top_n]
        ]

    return SearchBundleResponse(
        query=query,
        movie_details=details,
        tfidf_recommendations=tfidf_items,
        genre_recommendations=genre_recs,
    )
